+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - Added `Request.ValidateClientSide()` and `RequestSet.ValidateClientSide()`
 - Added `ResponseSet` and `UnmarshalResponseSet()`
 - Added `Response.Validate()`
+- Added `Response.UnmarshalRequestID()`
 - Added `Error.MarshalData()` and `UnmarshalData()`
 
 ### Changed